Refactor nested mutation handling
