+++ .changeset/few-poets-happen.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/keystone': patch
+---
+
+Refactored internal nested mutation input handler code.
